+++ packages/hooks/CHANGELOG.md
@@ -3,4 +3,5 @@
 ## v1.21.0
 
 - Removed babel
-- Features and services names are extensively derived by their registration form and automatically prefixed
+- Features and services names defaults are derived by their registration form and automatically prefixed
+- Added support for positional definition in `registerRoute`